==> COVID 19 Project Scripts/01. COVID Data Import and Cleaning- Stay at Home Orders.ipynb <==
# Contents
### 1. Import Libraries
### 2. Upload Data - Stay at Home Orders
### 3. Consistency Checks and Cleaning
* Drop Columns
* Find Duplicates
* Change Column Formats
* Remove Unwanted Values
* Reduce Memory by Changing Data Types
* Check for Missing Values
* Check for Mixed Types
* Re-Name Columns for Later Merge

### 4. Download Cleaned Data
_____________________

## 1. Import Libraries, Create Folder Path

# Import libraries
import pandas as pd
import numpy as np
import os

# Create path to folder
path = r'/Users/caitlin/iCloud/Caitlin/COVID Data/Original Data'

path

## 2. Upload Data - Stay at Home Orders

# Import data file
COVID_home_order = pd.read_csv(os.path.join(path, 'COVID_home_orders.csv'), index_col = False)

# Show all columns
pd.options.display.max_columns = None

# Examine shape
COVID_home_order.shape

# Check info
COVID_home_order.info()

# Examine columns for home order data
COVID_home_order.head()

## 3. Consistency Checks and Cleaning

### Drop Columns

# Drop unnecessary columns
## The source of the data is not important - the focus is on the actual close orders.
COVID_home_2 = COVID_home_order.drop(columns = ['Express_Preemption', 'Source_of_Action','URL','Citation'],)

# Check column drop
COVID_home_2.head()

### Find Duplicates

# Find home order duplicates
COVID_home_order_dups = COVID_home_2[COVID_home_2.duplicated()]

COVID_home_order_dups

COVID_home_order_dups.shape

No duplicates found.

### Change Column Formats

I am merging this dataset with mask mandate data. That dataset starts 10 April. I am also only interested in the time before vaccines because I do not want to adjust for vaccines as another variable. The first vaccine was issued in the US on 14 Dec 2020, so I am ending on that date.

# Change date column to datetime format so it behaves as a number
COVID_home_2['date'] = pd.to_datetime(COVID_home_2['date'])

# Check change
COVID_home_2.info()

### Remove Unwanted Values

# Check min and max values to see which ones to eliminate
COVID_home_2['date'].max()

COVID_home_2['date'].min()

# Make subset of desired time values
COVID_home_3 = COVID_home_2[(COVID_home_2['date'] >= '2020-04-10') & (COVID_home_2['date'] < '2020-12-14')]

COVID_home_3['date'].max()

COVID_home_3['date'].min()

COVID_home_3.shape

Only interested in 50 states and DC, eliminate other values.

# Group by to index state names by FIPS code
COVID_home_3.groupby(['FIPS_State'])['State_Tribe_Territory'].value_counts()

# Remove all but 50 states and DC
COVID_home_4 = COVID_home_3[(COVID_home_3.FIPS_State != 78) 
                            & (COVID_home_3.FIPS_State != 72) 
                            & (COVID_home_3.FIPS_State != 69)
                            & (COVID_home_3.FIPS_State != 66)
                            & (COVID_home_3.FIPS_State != 60)]

# Index by group by to confirm drop
COVID_home_4.groupby(['FIPS_State'])['State_Tribe_Territory'].value_counts(dropna = False)

COVID_home_4.shape

###  Reduce Memory by Changing Data Types

The DF is taking up a lot of memory, so changing int64 where there are not as many values.

# Change int values to lower int values to save memory - order code only has 2 values so int8
COVID_home_4['Order_code'] = COVID_home_4['Order_code'].astype('int8')

# Change int values to lower int values to save memory - FIPS only has ~50 values so int8
COVID_home_4['FIPS_State'] = COVID_home_4['FIPS_State'].astype('int8')

#Check max value for county for int change
COVID_home_4['FIPS_County'].max()

# Change FIPS county to int 16 because that has max value over 30,000
COVID_home_4['FIPS_County'] = COVID_home_4['FIPS_County'].astype('int16')

# Check final changes - significantly reduced memory
COVID_home_4.info()

### Check for Missing Values

# Check for missing values in home order dataframe
COVID_home_4.isnull().sum()

#Create null subset
COVID_home_nan = COVID_home_4[COVID_home_4['Stay_at_Home_Order_Recommendation'].isnull() == True]

COVID_home_nan

# Count values of nan values.
COVID_home_nan['Order_code'].value_counts(dropna=False)

The NaNs have been inaccurately labeled. They are no orders to stay home. Change to that code. They are all 7, which is stay at home, so that needs to be changed.

# Rename values from NaN to no order
COVID_home_4['Stay_at_Home_Order_Recommendation'] = COVID_home_4['Stay_at_Home_Order_Recommendation'].fillna('No order for individuals to stay home')

# Check values have been changed
COVID_home_4.isnull().sum()

### Check for Mixed Types

# Check for mixed types
for col in COVID_home_4.columns.tolist():
    weird = (COVID_home_4[[col]].applymap(type) != COVID_home_4[[col]].iloc[0].apply(type)).any(axis = 1) 
    if len (COVID_home_4[weird]) > 0:
        print (col)

It appears that through cleaning the data, mixed type was addressed.

#Check for Duplicates
COVID_home_dups = COVID_home_4[COVID_home_4.duplicated()]

COVID_home_dups

No duplicates

There is no numerical data for this dataset, so I am just cleaning and downloading it.

### Re-Name Columns for Later Merge

# Rename column for later merge
COVID_home_5 = COVID_home_4.rename(columns = {'Order_code' : 'order_code_home'})

COVID_home_5 = COVID_home_4.rename(columns = {'State_Tribe_Territory' : 'state'})

COVID_home_5.head()

## 4. Download Cleaned Data

#Download cleaned data
COVID_home_5.to_pickle(os.path.join(path, 'COVID_home_clean.pkl'))
