==> COVID 19 Project Scripts/02. COVID Data Import and Cleaning - Mask  Orders.ipynb <==
# Contents
### 1. Import Libraries
### 2. Upload Data - Mask Mandates
### 3. Consistency Checks and Cleaning
* Drop Columns
* Find Duplicates
* Change Data Types
* Remove Unwanted Values
* Reduce Memory by Changing Data Types
* Check for Missing Values
* Re-Name Columns for Later Merge

### 4. Download Cleaned Data
_______

## 1. Import Libraries, Create Folder Path

# Import libraries
import pandas as pd
import numpy as np
import os

# Create path to folder
path = r'/Users/caitlin/iCloud/Caitlin/COVID Data/Original Data'

path

## 2. Upload Data - Mask Mandates

# Download mask mandate data
COVID_mask_order = pd.read_csv(os.path.join(path, 'COVID_mask_mandates.csv'))

# Examine shape
COVID_mask_order.shape

# Examine columns for mask mandate data
COVID_mask_order.head()

## 3. Consistency Checks and Cleaning - Mask Mandates

# Check info
COVID_mask_order.info()

### Drop Columns

# Drop unnecessary columns
## The source of the data is not important - the focus is on the actual close orders.
COVID_mask_order_2 = COVID_mask_order.drop(columns = ['Source_of_Action','URL','Citation'],)

# Check column drop
COVID_mask_order_2.head()

### Find Duplicates

# Find mask mandate duplicates
COVID_mask_order_dups = COVID_mask_order_2[COVID_mask_order_2.duplicated()]

COVID_mask_order_dups

COVID_mask_order_dups.shape

No duplicates

### Change Data Types

# Change date column to datetime format so it behaves as a number
COVID_mask_order_2['date'] = pd.to_datetime(COVID_mask_order_2['date'])

COVID_mask_order_2.info()

### Remove Unwanted Values

# Check min and max values to see which ones to eliminate
COVID_mask_order_2['date'].max()

COVID_mask_order_2['date'].min()

I am only interested in the time before vaccines (14 Dec) so eliminating other values.

# Make subset of desired time values
COVID_mask_3 = COVID_mask_order_2[(COVID_mask_order_2['date'] < '2020-12-14')]

COVID_mask_3['date'].max()

# Group by to index state names by FIPS code
COVID_mask_3.groupby(['FIPS_State'])['State_Tribe_Territory'].value_counts()

# Remove all but 50 states and DC
COVID_mask_4 = COVID_mask_3[(COVID_mask_3.FIPS_State != 78) 
                            & (COVID_mask_3.FIPS_State != 72) 
                            & (COVID_mask_3.FIPS_State != 69)
                            & (COVID_mask_3.FIPS_State != 66)
                            & (COVID_mask_3.FIPS_State != 60)]

# Index by group by to confirm drop
COVID_mask_4.groupby(['FIPS_State'])['State_Tribe_Territory'].value_counts()

COVID_mask_4.shape

### Reduce Memory by Changing Data Types

# Change int values to lower int values to save memory - order code only has 2 values so int8
COVID_mask_4['order_code'] = COVID_mask_4['order_code'].astype('int8')

# Change int values to lower int values to save memory - FIPS only has ~50 values so int8
COVID_mask_4['FIPS_State'] = COVID_mask_4['FIPS_State'].astype('int8')

#Check max value for county for int change
COVID_mask_4['FIPS_County'].max()

# Change FIPS county to int 16 because that has max value over 30,000
COVID_mask_4['FIPS_County'] = COVID_mask_4['FIPS_County'].astype('int16')

# Check final changes - significantly reduced memory
COVID_mask_4.info()

### Check for Missing Values

# Check for missing values in mask mandate dataframe
COVID_mask_4.isnull().sum()

# Value counts to check source of NaNs
COVID_mask_4['Face_Masks_Required_in_Public'].value_counts(dropna=False)

# Group order code and mask recommendation to see how many of each
COVID_mask_4.groupby(['order_code'])['Face_Masks_Required_in_Public'].value_counts(dropna=False)

# Value counts to check source of NaNs
COVID_mask_4['order_code'].value_counts(dropna=False)

# Create dataframe for NaNs - masks
COVID_mask_nan = COVID_mask_4[COVID_mask_4['Face_Masks_Required_in_Public'].isnull() == True]

COVID_mask_nan['order_code'].value_counts(dropna=False)

All of the NaN values in the mask mandate database have the same code - 2 - which is that there is no mask mandate, as defined by the CDC. Since the 2s are equal to the NaN and NO, I am going to change the NaNs to reflect those values.

# Change NaN to No
COVID_mask_4['Face_Masks_Required_in_Public'] = COVID_mask_4['Face_Masks_Required_in_Public'].fillna('No')

# Group order code and mask recommendation to see how many of each
COVID_mask_4.groupby(['order_code'])['Face_Masks_Required_in_Public'].value_counts(dropna=False)

###  Re-Name Columns for Later Merge

# Changing column names
COVID_mask_4.rename(columns = {'State_Tribe_Territory' : 'state'}, inplace = True)

COVID_mask_4.rename(columns = {'order_code' : 'order_code_mask'}, inplace = True)

COVID_mask_4.head()

## 4. Download Cleaned Data

#Download mask data
COVID_mask_4.to_pickle(os.path.join(path, 'COVID_mask_clean.pkl'))
