==> COVID 19 Project Scripts/03. COVID Data Import and Cleaning - Cases Per 100K.ipynb <==
# Contents
### 1. Import Libraries
### 2. Upload Data - COVID Cases Per 100K
### 3. Consistency Checks and Cleaning
* Re-Name Columns
* Delete Unwanted Columns
* Change Column Formats
* Eliminate Unwanted Rows
* Change Data Types
* Check for Missing Values
* Check for Duplicates
* Check for Mixed Types
* Derive New Variables

### 4. Download Cleaned Data
_____________________

## 1. Import Libraries, Create Folder Path

# Import libraries
import pandas as pd
import numpy as np
import os

# Create path to folder
path = r'/Users/caitlin/iCloud/Caitlin/COVID Data/Original Data'

path

## 2. Upload Data - COVID Cases Per 100K

# Import data file
COVID_cases = pd.read_csv(os.path.join(path, 'COVID_cases.csv'), index_col = False)

# Show all columns
pd.options.display.max_columns = None

# Inspect Data File
COVID_cases.info()

# Inspect info
COVID_cases.info()

# Inspect shape
COVID_cases.shape

# Inspect DF heads
COVID_cases.head(5)

## 3. Consistency Checks and Cleaning

### Rename Columns

# Rename columns for clarity
COVID_cases_2 = COVID_cases.rename(columns = {'cases_per_100K_7_day_count_change' : 'cases_per_100K'})

# Check change
COVID_cases_2.head(3)

### Delete Unwanted Columns

# Drop unnecessary columns
COVID_cases_3 = COVID_cases_2.drop(columns = ['percent_test_results_reported_positive_last_7_days'])

# Inspect drop
COVID_cases_3.head(5)

#### Check that there are only states, no territories

I am only interested in the US states, not territories, so deleting like I did for the other two DFs.

# Count values of each state to see states and how many records for each
COVID_cases_3['state_name'].value_counts()

Only states, no territories - good. No deletions necessary.

### Change Column Formats

# Change date column to datetime format so it behaves as a number
COVID_cases_3['date'] = pd.to_datetime(COVID_cases_3['date'])

# Check change of data type
COVID_cases_3.info()

### Eliminate Unwanted Rows

I am only looking at data from 10 April (first available data on home and mask orders) to 13 December (before the first vaccine).

# Check min and max values to see which ones to eliminate
COVID_cases_3['date'].max()

# Check min and max values to see which ones to eliminate
COVID_cases_3['date'].min()

# Make subset of desired time values
COVID_cases_4 = COVID_cases_3[(COVID_cases_3['date'] >= '2020-04-10') & (COVID_cases_3['date'] < '2020-12-14')]

# Check min and max values again
COVID_cases_4['date'].max()

COVID_cases_4['date'].min()

### Change Data Types

# Change cases per 100K to numerical and coerce suppressed to NaN
COVID_cases_4['cases_per_100K'] = pd.to_numeric(COVID_cases_4['cases_per_100K'], errors='coerce')

# Check datatype change
COVID_cases_4.info()

# Min and max values to double check
COVID_cases_4['cases_per_100K'].min()

COVID_cases_4['cases_per_100K'].max()

### Check for Missing Values

# Check for missing values in home order dataframe
COVID_cases_4.isnull().sum()

There are a significant number NaN values in the cases per 100K, but I do not want to delete them because they are simply just days where the data was not reported, according to the CDC.

### Check for Duplicates

#Check for Duplicates
COVID_cases_dups = COVID_cases_4[COVID_cases_4.duplicated()]

COVID_cases_dups

No duplicates

### Check for Mixed Types

# Check for mixed types
for col in COVID_cases_4.columns.tolist():
    weird = (COVID_cases_4[[col]].applymap(type) != COVID_cases_4[[col]].iloc[0].apply(type)).any(axis = 1) 
    if len (COVID_cases_4[weird]) > 0:
        print (col)

No mixed types

#Download cleaned data
COVID_cases_4.to_csv(os.path.join(path, 'COVID_cases_per_100K.csv'))

### Derive New Variables

#### Add +14, +30, +60 Day Columns

A major premise of this project is determining how mask and stay at home mandates impacted COVID cases and deaths. Since COVID has a 14 day incubation period, a column that adds 14, 30, and 60 days to the date of the mandate is important to track the effectiveness of the mandates. A new DF will create the new variable, and then they will be merged.

# Import Timedelta
from datetime import timedelta

# Duplicate DF for later variable creation
COVID_cases_5 = COVID_cases_4

COVID_cases_5.head(5)

#### +14 Day Variable

# Create new column for +14 days
COVID_cases_5['plus_14_days'] = COVID_cases_5['date'] + timedelta(days=14)

# START HERE - FIGURE OUT HOW TO POPULATE THE +14 COLUMN with 100K VALUE from that date

COVID_cases_5.info()

COVID_cases_5.head(3)

df['value2'] = np.where(df['condition'].ge(0),0,df['value'])

COVID_cases_6['100K_+14'] = np.where(COVID_cases_5['date'] == COVID_cases_5['plus_14_days'].ge(0),0,COVID_cases_5['cases_per_100K'])

COVID_cases_6['100K_+14'] = np.where(COVID_cases_5['date'] == COVID_cases_5['plus_14_days'], COVID_cases_5['cases_per_100K'])

np.where(df["Number"] > 0.5, df["Country"], "")

COVID_cases_5['date'] == COVID_cases_5['plus_14.days']

COVID_cases_5['date'] == COVID_cases_5['plus_14.days'].values

master_df.loc[master_df['department_id'] == 1, 'department_name'] = "frozen"

# Using DataFrame.query() method extract column values.
COVID_cases_6=COVID_cases_5.query('Fee == 25000')['Courses']
print(df2)

#Output:
r2    Pyspark
Name: Courses, dtype: object

COVID_cases_6['cases_+14'] = COVID_cases_5.loc[(COVID_cases_5['date'] = COVID_cases_5['plus_14_days'])['cases_per_100K']

df2=df.loc[(df['Discount'] >= 1000) & (df['Discount'] <= 2000)]
print(df2)

COVID_cases_5.loc[(COVID_cases_5['date'] == COVID_cases_5['plus_14_days'].apply(max, axis=0)

COVID_cases_6 = COVID_cases_5.loc[(COVID_cases_5['date'] >= 1000) & (df['Discount'] <= 2000)]

COVID_cases_5['cases_+14'] = COVID_cases_5

COVID_cases_5.loc[COVID_cases_5['customer_number_dependents'] == 0, 'parental_status'] = "non-parent"

E_Price=[]
plus_14_cases = COVID_cases_5['timestamp'].apply(lambda x: x >= (COVID_cases_5['timestamp']+timedelta(minutes=2)))
price_diff = df['Price'].apply(lambda x: x >= (df['Price']*1.1))
for i in range(len(df)):
    check = (time_diff|price_diff)[i]
    ind = check.idxmax()
    if ind != 0:
        val = df.iloc[ind,1]
    else:
        val = np.nan
    E_Price.append(val)    

df['Exit_Price'] = E_Price
df['Exit_Price'] = df.Exit_Price.astype(pd.Int32Dtype())
print(df)

E_Price=[]
time_diff = df['timestamp'].apply(lambda x: x >= (df['timestamp']+timedelta(minutes=2)))
price_diff = df['Price'].apply(lambda x: x >= (df['Price']*1.1))
for i in range(len(df)):
    check = (time_diff|price_diff)[i]
    ind = check.idxmax()
    if ind != 0:
        val = df.iloc[ind,1]
    else:
        val = np.nan
    E_Price.append(val)    

df['Exit_Price'] = E_Price
df['Exit_Price'] = df.Exit_Price.astype(pd.Int32Dtype())
print(df)

# Delete columns on new DF to facilitate merge
COVID_cases_6 = COVID_cases_5.drop(columns = ['date','cases_per_100K','community_transmission_level'],)

COVID_cases_6.head(3)

# Re-name date column
COVID_cases_5 = COVID_cases_5.rename(columns = {'plus_14_days' : 'date', 'cases_per_100K' : 'cases_100K_+14'})

COVID_cases_5.head(3)

COVID_cases_4.head(3)

COVID_cases_4 = COVID_cases_4.drop(columns = ['plus_14_days'])

COVID_cases_4.head(3)

# Merge dataframes to create new variable
COVID_cases_6 = COVID_cases_4.merge(COVID_cases_5, on = ['date', 'state_name','county_name','fips_code','community_transmission_level'])

COVID_cases_6.head(3)

#### +30 Day Variable

# Create new column for +30 days
COVID_cases_6["plus_30_days"] = COVID_cases_6["date"] + timedelta(days=30)

COVID_cases_6.head(3)

# Delete date on new DF to facilitate merge
COVID_cases_7 = COVID_cases_6.drop(columns = ['date'])

COVID_cases_7.head(3)

# Re-name date column
COVID_cases_8 = COVID_cases_7.rename(columns = {'plus_30_days' : 'date', 'cases_per_100K' : 'cases_100K_+30'})

COVID_cases_8.head(3)

COVID_cases_6.head(3)

COVID_cases_6 = COVID_cases_6.drop(columns = ['plus_30_days'])

COVID_cases_6.head(3)

# Merge dataframes to create new variable
COVID_cases_9 = COVID_cases_6.merge(COVID_cases_8, on = ['date', 'state_name','county_name','fips_code','community_transmission_level','cases_100K_+14'])

COVID_cases_9.head(3)

#### + 60 Day Variable

# Create new column for +60 days
COVID_cases_9["plus_60_days"] = COVID_cases_9["date"] + timedelta(days=60)

COVID_cases_9.head(3)

# Delete date on new DF to facilitate merge
COVID_cases_10 = COVID_cases_9.drop(columns = ['date'])

# Re-name date column
COVID_cases_11 = COVID_cases_10.rename(columns = {'plus_60_days' : 'date', 'cases_per_100K' : 'cases_100K_+60'})

COVID_cases_11.head(3)

COVID_cases_9.head(3)

COVID_cases_9 = COVID_cases_9.drop(columns = ['plus_60_days'])

# Merge dataframes to create new variable
COVID_cases_12 = COVID_cases_9.merge(COVID_cases_11, on = ['date', 'state_name','county_name','fips_code','community_transmission_level','cases_100K_+14','cases_100K_+30'])

COVID_cases_12.head(3)

## 4. Download Cleaned Data

#Download cleaned data
COVID_cases_12.to_csv(os.path.join(path, 'COVID_cases_per_100K.csv'))
