==> COVID 19 Project Scripts/04. COVID Data Import and Cleaning - COVID Cases and Deaths.ipynb <==
# Contents
### 1. Import Libraries
### 2. Upload Data - COVID Cases and Deaths
### 3. Consistency Checks and Cleaning
* Drop Columns
* Find Duplicates
* Change Column Formats
* Remove Unwanted Values
* Check for Missing Values
* Check for Mixed Types

### 4. Download Cleaned Data
_____________________

## 1. Import Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# Create path to folder
path = r'/Users/caitlin/iCloud/Caitlin/COVID Data/Original Data'

path

## 2. Upload Data - COVID Cases and Deaths

# Import data file
COVID_cases = pd.read_csv(os.path.join(path, 'COVID_cases_deaths.csv'), index_col = False)

# Show all columns
pd.options.display.max_columns = None

# Examine shape
COVID_cases.shape

# Check info
COVID_cases.info()

# Examine columns for covid cases
COVID_cases.head()

## 3. Consistency Checks and Cleaning 

### Drop Columns

# Drop unnecessary columns
## Only using total for each, not presumed or probable.
COVID_cases_2 = COVID_cases.drop(columns = ['pnew_case','conf_cases','conf_death','created_at','prob_cases','prob_death', 'pnew_death', 'consent_cases', 'consent_deaths'])

COVID_cases_2.head()

### Find Duplicates

# Find duplicates
COVID_cases_dups = COVID_cases_2[COVID_cases_2.duplicated()]

COVID_cases_dups

No duplicates

### Change Column Formats

# Change date column to datetime format so it behaves as a number
COVID_cases_2['submission_date'] = pd.to_datetime(COVID_cases_2['submission_date'])

COVID_cases_2.info()

# Rename date column for later merge
COVID_cases_3 = COVID_cases_2.rename(columns = {'submission_date' : 'date',
                                               'tot_cases' : 'Total Cases',
                                               'new_case' : 'New Cases',
                                               'tot_death' : 'Total Deaths',
                                               'new_death' : 'New Deaths'})

COVID_cases_3.head(2)

### Remove Unwanted Values

# Check min and max values to see which ones to eliminate
COVID_cases_3['date'].max()

# Check min and max values to see which ones to eliminate
COVID_cases_3['date'].min()

I am only interested in the time before the vaccine was administered - 14 December 2020. I will include 14 days after, however, because that is the incubation period for COVID.

# Make subset of desired date values
COVID_cases_3 = COVID_cases_3[(COVID_cases_3['date'] >= '2020-04-10') & (COVID_cases_3['date'] <= '2020-12-28')]

# Check min and max values to check subset
COVID_cases_3['date'].max()

COVID_cases_3['date'].min()

# Value counts to see states included
COVID_cases_3['state'].value_counts(dropna = False)

Only interested in 50 states and DC, take out others. 

COVID_cases_3 = COVID_cases_3[(COVID_cases_3.state != "RMI") & (COVID_cases_3.state != "FSM") & (COVID_cases_3.state != "GU") & (COVID_cases_3.state != "PW") & (COVID_cases_3.state != "VI") & (COVID_cases_3.state != "PR") & (COVID_cases_3.state != "MP")]

# Check omissions
COVID_cases_3['state'].value_counts()

### Check for Missing Values

# Check for missing values in home order dataframe
COVID_cases_3.isnull().sum()

No missing values.

### Check for Mixed Types

# Check for mixed types
for col in COVID_cases_3.columns.tolist():
    weird = (COVID_cases_3[[col]].applymap(type) != COVID_cases_3[[col]].iloc[0].apply(type)).any(axis = 1) 
    if len (COVID_cases_3[weird]) > 0:
        print (col)

No mixed types.

## 4. Download Cleaned Data

#Download cleaned data
COVID_cases_3.to_csv(os.path.join(path, 'COVID_cases_clean.csv'))
