==> COVID 19 Project Scripts/05. COVID Merging Datasets.ipynb <==
# Contents

### 1. Import Library 
### 2. Import  Mask and Stay and Home Data
### 3. Merge Stay at Home and Mask Data
### 4. Prepare Mask and Home Data for Merge
* Derive New Categorical Variables
* Add Variables for Dataset Agreement

### 5. Import COVID Data: Deaths Per 100K
### 6. Prepare COVID Data for Merge
*  Rename Columns

### 7. Merge Mask/Home with Cases Per 100K
### 8. Download Merged 100K and Mask/Home Order
### 9. Create Sub-Set for Merge with COVID Case and Death Data
* Derive New Grouped Variables
* Merge Subdata Sets
* Replace Values

### 10. Download COVID Case and Death Data
### 11. Prepare COVID Case and Death Data for Merge
* Drop Columns
* Rename Columns for Merge
* Change Data Format

### 12. Merge COVID Mask/Home/Case Per 100K to Cases and Death Data
###  13. Download Merged COVID Mask/Home/Case Per 100K to Cases and Death Data

## 1. Import Library and Data

# Import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

## 2. Import Mask and Stay at Home Data

# Create path to folder
path = r'/Users/caitlin/iCloud/Caitlin/COVID Data/Prepared Data'

path

#### Stay at home order data

# Import stay at home data file
COVID_home = pd.read_pickle(os.path.join(path, 'COVID_home_clean.pkl'))

COVID_home.head(2)

COVID_home.shape

#### Mask mandate data

# Import mask mandate data file
COVID_mask = pd.read_pickle(os.path.join(path, 'COVID_mask_clean.pkl'))

COVID_mask.head(2)

COVID_mask.shape

## 3. Merge Stay at Home and Mask Data

# Merge dataframes
COVID_mask_home = COVID_home.merge(COVID_mask, on = ['state', 'County_Name', 'FIPS_State', 'FIPS_County', 'date'])

COVID_mask_home.head(5)

COVID_mask_home.shape

COVID_mask_home.info()

## 4. Prepare Mask and Home Data for Merge

### Derive New Categorical Variables

A yes/no/recommended categorical variable will be analyzed later, so adding.

#### Stay at Home Orders

# Group order code and stay at home recommendation to see how many of each
COVID_mask_home.groupby(['order_code_home'])['Stay_at_Home_Order_Recommendation'].value_counts()

# Create exclusion flag for dates with no stay at home orders 
COVID_mask_home.loc[COVID_mask_home['order_code_home'] == 7, 'home_order'] = "No"
COVID_mask_home.loc[COVID_mask_home['order_code_home'] <= 5, 'home_order'] = "Yes"
COVID_mask_home.loc[COVID_mask_home['order_code_home'] == 6, 'home_order'] = "Recommended"

# Check creation of flag and that it equals 7 above
COVID_mask_home['home_order'].value_counts()

#### Mask Mandates

# Group order code and mask recommendation to see how many of each
COVID_mask_home.groupby(['order_code_mask'])['Face_Masks_Required_in_Public'].value_counts()

# Show all rows
pd.options.display.max_rows = None

# Check new values
COVID_mask_home.head(3)

### Add Variables for Dataset Agreement

Have to add a column for state names since the COVID case dataset has names.

# Rename current state abbreviation state column
COVID_mask_home = COVID_mask_home.rename(columns = {'state':'state_abbreviated'})

# Create dictionary for state abbreviations
us_state_abbrev = {
        'AK': 'Alaska',
        'AL': 'Alabama',
        'AR': 'Arkansas',
        'AZ': 'Arizona',
        'CA': 'California',
        'CO': 'Colorado',
        'CT': 'Connecticut',
        'DC': 'District of Columbia',
        'DE': 'Delaware',
        'FL': 'Florida',
        'GA': 'Georgia',
        'HI': 'Hawaii',
        'IA': 'Iowa',
        'ID': 'Idaho',
        'IL': 'Illinois',
        'IN': 'Indiana',
        'KS': 'Kansas',
        'KY': 'Kentucky',
        'LA': 'Louisiana',
        'MA': 'Massachusetts',
        'MD': 'Maryland',
        'ME': 'Maine',
        'MI': 'Michigan',
        'MN': 'Minnesota',
        'MO': 'Missouri',
        'MS': 'Mississippi',
        'MT': 'Montana',
        'NA': 'National',
        'NC': 'North Carolina',
        'ND': 'North Dakota',
        'NE': 'Nebraska',
        'NH': 'New Hampshire',
        'NJ': 'New Jersey',
        'NM': 'New Mexico',
        'NV': 'Nevada',
        'NY': 'New York',
        'OH': 'Ohio',
        'OK': 'Oklahoma',
        'OR': 'Oregon',
        'PA': 'Pennsylvania',
        'RI': 'Rhode Island',
        'SC': 'South Carolina',
        'SD': 'South Dakota',
        'TN': 'Tennessee',
        'TX': 'Texas',
        'UT': 'Utah',
        'VA': 'Virginia',
        'VT': 'Vermont',
        'WA': 'Washington',
        'WI': 'Wisconsin',
        'WV': 'West Virginia',
        'WY': 'Wyoming'}

# Create new column with full state names
COVID_mask_home['state'] = COVID_mask_home['state_abbreviated'].map(us_state_abbrev).fillna(COVID_mask_home['state_abbreviated'])

COVID_mask_home.head(5)

# Check addition
COVID_mask_home.head(5)

# Check counts to check add
COVID_mask_home['state'].value_counts(dropna = False)

## 5. Import COVID Data: Deaths Per 100K

# Import stay at COVID cases
COVID_cases = pd.read_csv(os.path.join(path, 'COVID_cases_per_100K.csv'))

COVID_cases.head(5)

# Delete unneeded columns
COVID_cases = COVID_cases.drop(columns = ['Unnamed: 0'],)

COVID_cases.head(3)

COVID_cases.info()

COVID_cases.shape

# Change date column to datetime format so it behaves as a number
COVID_cases['date'] = pd.to_datetime(COVID_cases['date'])

COVID_cases.info()

## 6. Prepare COVID Data for Merge

### Rename Columns

# Rename Columns
COVID_cases_2 = COVID_cases.rename(columns = {'state_name':'state','county_name':'County_Name'})

COVID_cases_2.head(3)

## 7. Merge Mask/Home with Cases

# Merge dataframes
COVID_mask_home_100K = COVID_mask_home.merge(COVID_cases_2, on = ['state', 'County_Name','date'])

COVID_mask_home_100K.head(3)

COVID_mask_home_100K.shape

## 8. Download Merged 100K and Mask/Home Order

The county level data could be useful, but in order to merge all the data, it requires combining to states as the key, so I am going to use two different data sets.

#Download cleaned data
COVID_mask_home_100K.to_csv(os.path.join(path, 'COVID_state_mask_100K.csv'))

COVID_mask_home_100K.groupby(['order_code_home'])['Stay_at_Home_Order_Recommendation'].value_counts()

## 9. Create Sub-Set for Merge with COVID Case and Death Data

Datasets above are on the county level, so variables need to be grouped by state to merge with case and death data.

### Derive New Grouped Variables

# Variable creation: Count of yes for stay at home recommended
COVID_home_by_state_yes = COVID_mask_home_100K.groupby(['date','state'])['home_order'].apply(lambda x: (x=='Yes').sum()).reset_index(name='home_yes_count')

# Variable creation: Count of no for stay at home recommended
COVID_home_by_state_no = COVID_mask_home_100K.groupby(['date','state'])['home_order'].apply(lambda x: (x=='No').sum()).reset_index(name='home_no_count')

# Variable creation: Count of recommend for stay at home recommended
COVID_home_by_state_recommend = COVID_mask_home_100K.groupby(['date','state'])['home_order'].apply(lambda x: (x=='Recommended').sum()).reset_index(name='home_recommend_count')

# Variable creation: Count of yes mask orders
COVID_mask_by_state_yes = COVID_mask_home_100K.groupby(['date','state'])['Face_Masks_Required_in_Public'].apply(lambda x: (x=='Yes').sum()).reset_index(name='yes_mask_count')

# Variable creation: Count of no mask orders
COVID_mask_by_state_no = COVID_mask_home_100K.groupby(['date','state'])['Face_Masks_Required_in_Public'].apply(lambda x: (x=='No').sum()).reset_index(name='no_mask_count')

# Variable creation: Sum of COVID cases by 100K
COVID_home_by_state_avgcases100K = COVID_mask_home_100K.groupby(['date','state'])['cases_per_100K'].mean().reset_index(name='avg_per_100K')

### Merge Subdata Sets

# Merge dataframes
COVID_home_by_state_total = COVID_home_by_state_yes.merge(COVID_home_by_state_no, on = ['date', 'state']).merge(COVID_home_by_state_recommend, on = ['date', 'state']).merge(COVID_mask_by_state_yes, on = ['date', 'state']).merge(COVID_mask_by_state_no, on = ['date', 'state']).merge(COVID_home_by_state_avgcases100K, on = ['date', 'state'])

# Check merge
COVID_home_by_state_total.head(5)

### Replace Values

Since counts are important to this analysis, and Python will count 0 as a value, I need to turn all 0 into NaNs for categorical variables.

# Replace 0 with NaNs
COVID_home_by_state_total['home_yes_count']=COVID_home_by_state_total['home_yes_count'].replace(0, np.nan)

COVID_home_by_state_total['home_no_count']=COVID_home_by_state_total['home_no_count'].replace(0, np.nan)

COVID_home_by_state_total['home_recommend_count']=COVID_home_by_state_total['home_recommend_count'].replace(0, np.nan)

COVID_home_by_state_total['yes_mask_count']=COVID_home_by_state_total['yes_mask_count'].replace(0, np.nan)

COVID_home_by_state_total['no_mask_count']=COVID_home_by_state_total['no_mask_count'].replace(0, np.nan)

COVID_home_by_state_total.head(20)

## 10. Download COVID Case and Death Data

# Import stay at home data file
COVID_cases_deaths = pd.read_csv(os.path.join(path, 'COVID_final_updated_2.csv'))

COVID_cases_deaths.head(10)

## 11. Prepare COVID Case and Death Data for Merge

### Drop Columns

# Delete unneeded columns
COVID_cases_deaths_2 = COVID_cases_deaths.drop(columns = ['Unnamed: 0','Unnamed: 0.1'],)

COVID_cases_deaths_2.head(3)

### Rename Columns for Merge

# Rename current state abbreviation state column
COVID_cases_deaths_3 = COVID_cases_deaths_2.rename(columns = {'state':'state_abbreviated'})

# Create dictionary for state abbreviations
us_state_abbrev = {
        'AK': 'Alaska',
        'AL': 'Alabama',
        'AR': 'Arkansas',
        'AZ': 'Arizona',
        'CA': 'California',
        'CO': 'Colorado',
        'CT': 'Connecticut',
        'DC': 'District of Columbia',
        'DE': 'Delaware',
        'FL': 'Florida',
        'GA': 'Georgia',
        'HI': 'Hawaii',
        'IA': 'Iowa',
        'ID': 'Idaho',
        'IL': 'Illinois',
        'IN': 'Indiana',
        'KS': 'Kansas',
        'KY': 'Kentucky',
        'LA': 'Louisiana',
        'MA': 'Massachusetts',
        'MD': 'Maryland',
        'ME': 'Maine',
        'MI': 'Michigan',
        'MN': 'Minnesota',
        'MO': 'Missouri',
        'MS': 'Mississippi',
        'MT': 'Montana',
        'NA': 'National',
        'NC': 'North Carolina',
        'ND': 'North Dakota',
        'NE': 'Nebraska',
        'NH': 'New Hampshire',
        'NJ': 'New Jersey',
        'NM': 'New Mexico',
        'NV': 'Nevada',
        'NY': 'New York',
        'OH': 'Ohio',
        'OK': 'Oklahoma',
        'OR': 'Oregon',
        'PA': 'Pennsylvania',
        'RI': 'Rhode Island',
        'SC': 'South Carolina',
        'SD': 'South Dakota',
        'TN': 'Tennessee',
        'TX': 'Texas',
        'UT': 'Utah',
        'VA': 'Virginia',
        'VT': 'Vermont',
        'WA': 'Washington',
        'WI': 'Wisconsin',
        'WV': 'West Virginia',
        'WY': 'Wyoming'}

# Create new column with full state names
COVID_cases_deaths_3['state'] = COVID_cases_deaths_3['state_abbreviated'].map(us_state_abbrev).fillna(COVID_cases_deaths_3['state_abbreviated'])

COVID_cases_deaths_3.head(5)

COVID_cases_deaths_3.info()

### Change Data Format

# Change date column to datetime format so it behaves as a number
COVID_cases_deaths_3['date'] = pd.to_datetime(COVID_cases_deaths_3['date'])

##  12. Merge COVID Mask/Home/Case Per 100K to Cases and Death Data

# Merge dataframes
COVID_by_state_final_merged = COVID_home_by_state_total.merge(COVID_cases_deaths_3, on = ['date', 'state'])

COVID_by_state_final_merged.head(5)

# Drop unnecessary columns
COVID_by_state_final_merged = COVID_by_state_final_merged.drop(columns = ['state_abbreviated'],)

COVID_by_state_final_merged.head(5)

## 13. Download Merged COVID Mask/Home/Case Per 100K to Cases and Death Data

#Download cleaned data
COVID_by_state_final_merged.to_csv(os.path.join(path, 'COVID_death_case_mandate_by_state.csv'))
